backend/requirements.txt
fastapi
uvicorn
python-multipart
pillow
pypdfium2
reportlab
pytesseract
